pyproject.toml
[project]
name = "plexio"
readme = "README.md"
requires-python = ">=3.11"
version = "0.0.1"
dependencies = [
    "fastapi==0.104.1",
    "pydantic-settings==2.0.1",
    "yarl==1.9.4",
    "uvicorn==0.24.0",
    "aiohttp==3.9.1",
    "redis==5.0.1",
    "sentry-sdk[fastapi]==1.38.0"
]

[project.optional-dependencies]
dev = [
    "ruff==0.1.7"
]

[tool.setuptools]
packages = ["plexio"]

[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[tool.ruff]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "C",   # flake8-comprehensions
    "B",   # flake8-bugbear
]
line-length = 88
target-version = "py311"

[tool.ruff.isort]
split-on-trailing-comma = true
force-wrap-aliases = true
combine-as-imports = true

[tool.ruff.format]
quote-style="single"
